// extensions/erdos-database-client/src/model/interface/sshConfig.ts
export interface SSHConfig {
    /**
     * local tunnel port
     */
    tunnelPort: number;
    host: string;
    port: number;
    username: string;
    password?: string;
    /**
     * password privateKey native
     */
    type?: string;
    privateKeyPath?: string;
    watingTime?: number;
    /**
     * only support private keys generated by ssh-keygen, which means pkcs8 is not support.
     */
    privateKey?: Buffer;
    passphrase?: string;
    algorithms?: Algorithms;
    /**
     * only ssh connection
     */
    key: string;
}

export interface Algorithms {
    cipher?: string[];
}